requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
plotly>=5.15.0
scikit-image>=0.21.0
networkx>=3.1
pandas>=2.0.0
scipy>=1.11.0